fixed powerupbar bug.  changed around case numbering in pattern selection to reflect new difficulties
git-svn-id: https://prok-software-design.googlecode.com/svn/trunk@43 14bf4150-d967-81f9-8f7b-c011daa192b1

+++ src/game/Player.java
@@ -25,12 +25,12 @@ public class Player {
 	private long time;
 	private long delay=100;
 	private Image outline = GameplayState.images[14];
-	public int powerUpTimer = 0;
+	public float powerUpTimer = 0;
 	public int currPowerUp = -1;
 	public boolean dd = false;
 	public boolean invul = false;
 	public boolean twarp = false;
-	public int powerTime=1;
+	public float powerTime=1;
 	public Color colorBar;
 
 	public Player(Point p){
